added authorization to vet routes

+++ controllers/veterinarians_controller.py
@@ -1,5 +1,5 @@
 from types import NoneType
-from flask import Blueprint, request
+from flask import Blueprint, request, abort
 from init import db, bcrypt
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from models.veterinarian import VeterinarianSchema, Veterinarian
@@ -12,33 +12,75 @@ veterinarians_bp = Blueprint('veterinarians', __name__, url_prefix='/veterinaria
 @veterinarians_bp.route('/')
 def get_all_veterinarians():
     veterinarians = gb.filter_all_records(Veterinarian)
-    return VeterinarianSchema(many=True, exclude=['password']).dump(veterinarians)
+    return VeterinarianSchema(many=True, only=['first_name', 'last_name', 'description', 'email', 'sex', 'languages']).dump(veterinarians)
 
+@veterinarians_bp.route('/full_details/')
+@jwt_required()
+def get_all_veterinarians_full_details():
+    if gb.is_admin():
+        veterinarians = gb.filter_all_records(Veterinarian)
+        return VeterinarianSchema(many=True, exclude=['password']).dump(veterinarians)
+    else:
+        abort(401)
+
+@veterinarians_bp.route('/appointments/')
+@jwt_required()
+def get_all_appointments():
+    if gb.is_admin():
+        veterinarians = gb.filter_all_records(Veterinarian)
+        return VeterinarianSchema(many=True, only=['appointments', 'id']).dump(veterinarians)
+    else:
+        abort(401)
 
 @veterinarians_bp.route('/<int:veterinarian_id>/')
 def get_one_veterinarian(veterinarian_id):
     veterinarian = gb.required_record(Veterinarian, veterinarian_id)
-    return VeterinarianSchema(exclude=['password']).dump(veterinarian)
+    return VeterinarianSchema(only=['first_name', 'last_name', 'description', 'email', 'sex', 'languages']).dump(veterinarian)
 
+@veterinarians_bp.route('/<int:veterinarian_id>/full_details/')
+@jwt_required()
+def get_one_veterinarian_full_details(veterinarian_id):
+    if gb.is_admin() or gb.is_authorized_veterinarian(veterinarian_id):
+        veterinarian = gb.required_record(Veterinarian, veterinarian_id)
+        return VeterinarianSchema(exclude=['password']).dump(veterinarian)
+    else:
+        abort(401)
+
+@veterinarians_bp.route('/<int:veterinarian_id>/appointments/')
+@jwt_required()
+def get_one_veterinarian_appointments(veterinarian_id):
+    if gb.is_admin() or gb.is_authorized_veterinarian(veterinarian_id):
+        veterinarian = gb.required_record(Veterinarian, veterinarian_id)
+        return VeterinarianSchema(only=['appointments']).dump(veterinarian)
+    else:
+        abort(401)
 
 @veterinarians_bp.route('/<int:veterinarian_id>/', methods=['DELETE'])
+@jwt_required()
 def delete_veterinarian(veterinarian_id):
-    veterinarian = gb.required_record(Veterinarian, veterinarian_id)
-    db.session.delete(veterinarian)
-    db.session.commit()
-    return {'msg': f'Veterinarian {veterinarian.first_name} {veterinarian.last_name} deleted successfully'}
+    if gb.is_admin():
+        veterinarian = gb.required_record(Veterinarian, veterinarian_id)
+        db.session.delete(veterinarian)
+        db.session.commit()
+        return {'msg': f'Veterinarian {veterinarian.first_name} {veterinarian.last_name} deleted successfully'}
+    else:
+        abort(401)
 
 
 @veterinarians_bp.route('/<int:veterinarian_id>/', methods=['PUT', 'PATCH'])
+@jwt_required()
 def update_veterinarian(veterinarian_id):
-    veterinarian = gb.required_record(Veterinarian, veterinarian_id)
-    for key in list(request.json.keys()):
-        if key in ['languages', 'description']:
-            setattr(veterinarian, key, gb.nullable_value_converter(veterinarian, key))
-        else:
-            setattr(veterinarian, key, gb.required_value_converter(veterinarian, key))
-    db.session.commit()
-    return VeterinarianSchema(exclude=['password']).dump(veterinarian)
+    if gb.is_admin() or gb.is_authorized_veterinarian(veterinarian_id):
+        veterinarian = gb.required_record(Veterinarian, veterinarian_id)
+        for key in list(request.json.keys()):
+            if key in ['languages', 'description']:
+                setattr(veterinarian, key, gb.nullable_value_converter(veterinarian, key))
+            else:
+                setattr(veterinarian, key, gb.required_value_converter(veterinarian, key))
+        db.session.commit()
+        return VeterinarianSchema(exclude=['password']).dump(veterinarian)
+    else:
+        abort(401)
 
 
 @veterinarians_bp.route('/register/', methods=['POST'])
